fix(autoware_ground_segmentation): fix scan ground filter logic  (autowarefoundation/autoware_universe#9084)
* refactor: initialize gnd_grids in ScanGroundFilterComponent::initializeFirstGndGrids

Initialize gnd_grids vector in the ScanGroundFilterComponent::initializeFirstGndGrids function to ensure it is empty and has the correct capacity. This improves the efficiency of the function and ensures accurate grid initialization.

Signed-off-by: Taekjin LEE <taekjin.lee@tier4.jp>

* refactor: initialize gnd_grids vector in initializeFirstGndGrids function

Initialize the gnd_grids vector in the initializeFirstGndGrids function to ensure it is empty and has the correct capacity. This improves the efficiency of the function and ensures accurate grid initialization.

Signed-off-by: Taekjin LEE <taekjin.lee@tier4.jp>

* refactor: improve efficiency and accuracy of grid initialization

Initialize the gnd_grids vector in the initializeFirstGndGrids function to ensure it is empty and has the correct capacity. This refactor improves the efficiency of the function and ensures accurate grid initialization.

Signed-off-by: Taekjin LEE <taekjin.lee@tier4.jp>

* refactor: improve efficiency of checkDiscontinuousGndGrid function

Refactor the checkDiscontinuousGndGrid function in node.cpp to improve its efficiency. The changes include optimizing the conditional statements and reducing unnecessary calculations.

Signed-off-by: Taekjin LEE <taekjin.lee@tier4.jp>

* refactor: improve efficiency of checkDiscontinuousGndGrid function

Signed-off-by: Taekjin LEE <taekjin.lee@tier4.jp>

* fix: add missing condition

Signed-off-by: Taekjin LEE <taekjin.lee@tier4.jp>

* style(pre-commit): autofix

* refactor: fix height_max initialization in node.hpp

Signed-off-by: Taekjin LEE <taekjin.lee@tier4.jp>

* fix: bring back inequality sign

Signed-off-by: Taekjin LEE <taekjin.lee@tier4.jp>

* fix: parameters from float to double

following the guideline https://docs.ros.org/en/foxy/Concepts/About-ROS-2-Parameters.html#overview

Signed-off-by: Taekjin LEE <taekjin.lee@tier4.jp>

* refactor: fix logic description comment

Signed-off-by: Taekjin LEE <taekjin.lee@tier4.jp>

---------

Signed-off-by: Taekjin LEE <taekjin.lee@tier4.jp>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>